[Basic tuto 3: dynamic pipelines] Tuto + doc.
Usage:
./basic-3_dynamic_pipelines <path>

Reads a file with uridecodebin element and waits for pad-added signal.

When received, it checks the pad type and if it's “audio/x-raw”,
it attempts to link it with the audioconvert element.

This is a more complicated way to read a stream than basic-1, but
this way allows to choose the part of the stream to read.